auth: append random suffix when derive username from email
