+++ cmd/frontend/auth/BUILD.bazel
@@ -58,6 +58,7 @@ go_test(
         "//cmd/frontend/internal/app/ui",
         "//internal/actor",
         "//internal/auth",
+        "//internal/auth/userpasswd",
         "//internal/conf",
         "//internal/database",
         "//internal/database/dbmocks",
