+++ internal/scim/BUILD.bazel
@@ -23,6 +23,7 @@ go_library(
     deps = [
         "//cmd/frontend/auth",
         "//cmd/frontend/globals",
+        "//internal/auth/userpasswd",
         "//internal/authz",
         "//internal/conf",
         "//internal/database",
